Implement group of permissions
Permissions were ordered by property keys which is not guaranteed in JavaScript, therefore some related permissions were completely separated. Permissions are now stored in array which gives the possibility to specify a precise order, additionnaly it is possible to order permissions by groups. Top level permissions not inside a group are automatically put in a generic group.